requirements.txt
numpy>=1.24,<1.27
Pillow>=10.0
opencv-python-headless>=4.8
scikit-image>=0.21
google-genai>=1.0.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
tenacity>=8.1.0,<9
anyio>=3.7.1
FuzzyTM>=0.4.0

# HTTP (navigation / Maps)
requests>=2.31.0

# WebSocket server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
